+++ R/imports.R
@@ -6,9 +6,12 @@
 #' @importFrom dplyr %>% arrange bind_rows filter group_by if_else mutate rename rowwise select summarize ungroup
 #' @importFrom stringr str_extract str_remove str_replace str_split
 #' @importFrom tidyr pivot_longer pivot_wider replace_na
-#' @importFrom stats IQR as.formula model.matrix quantile
+#' @importFrom stats IQR as.formula model.matrix quantile setNames
 #' @importFrom rlang .data :=
 #' @importFrom utils txtProgressBar setTxtProgressBar
+#' @import jsonlite
+#' @import rxnorm
+#' @import httr
 #'
 #' @keywords internal
 NULL